refactoring: type 분리
